requirements.txt
flask==2.3.3
requests==2.32.3
Pillow==10.0.0
flask-cors==4.0.0
transformers>=4.30.0
torch>=2.0.0